refactor: use non-breaking wrapper
