+++ packages/repack/src/plugins/ModuleFederationPlugin.ts
@@ -1,19 +1,19 @@
-import type { moduleFederationPlugin as MF } from '@module-federation/sdk';
 import type { Compiler, RspackPluginInstance } from '@rspack/core';
-import { isRspackCompiler } from './utils/isRspackCompiler';
+// biome-ignore lint/correctness/noUnusedImports: needed for jsdoc
+import type { Federated } from '../utils/federated';
+import {
+  ModuleFederationPluginV1,
+  type ModuleFederationPluginV1Config,
+} from './ModuleFederationPluginV1';
 
 /**
- * {@link ModuleFederationPlugin} configuration options.
+ * {@link ModuleFederationPluginV1Config} configuration options.
  *
- * The fields and types are exactly the same as in the official `ModuleFederationPlugin`.
+ * The fields and types are exactly the same as in `webpack.container.ModuleFederationPlugin`.
  *
- * You can check documentation for all supported options here: https://module-federation.io/configure/
+ * You can check documentation for all supported options here: https://webpack.js.org/plugins/module-federation-plugin/
  */
-export interface ModuleFederationPluginConfig
-  extends MF.ModuleFederationPluginOptions {
-  /** Enable or disable adding React Native deep imports to shared dependencies */
-  reactNativeDeepImports?: boolean;
-}
+export type ModuleFederationPluginConfig = ModuleFederationPluginV1Config;
 
 /**
  * Webpack plugin to configure Module Federation with platform differences
@@ -81,183 +81,24 @@ export interface ModuleFederationPluginConfig
  * @category Webpack Plugin
  */
 export class ModuleFederationPlugin implements RspackPluginInstance {
-  private config: MF.ModuleFederationPluginOptions;
-  private deepImports: boolean;
+  private plugin: ModuleFederationPluginV1;
 
-  constructor(pluginConfig: ModuleFederationPluginConfig) {
-    const { reactNativeDeepImports, ...config } = pluginConfig;
-    this.config = config;
-    this.deepImports = reactNativeDeepImports ?? true;
-  }
-
-  private ensureModuleFederationPackageInstalled(context: string) {
-    try {
-      require.resolve('@module-federation/enhanced', { paths: [context] });
-    } catch {
-      throw new Error(
-        "ModuleFederationPlugin requires '@module-federation/enhanced' to be present in your project. " +
-          'Did you forget to install it?'
-      );
-    }
-  }
-
-  private adaptRuntimePlugins(
-    context: string,
-    runtimePlugins: string[] | undefined = []
-  ) {
-    const repackRuntimePlugin = require.resolve(
-      '../modules/FederationRuntimePlugin'
-    );
-
-    const plugins = runtimePlugins
-      .map((pluginPath) => {
-        try {
-          // resolve the paths to compare against absolute paths
-          return require.resolve(pluginPath, { paths: [context] });
-        } catch {
-          // ignore invalid paths
-          return undefined;
-        }
-      })
-      .filter((pluginPath) => !!pluginPath) as string[];
-
-    if (!plugins.includes(repackRuntimePlugin)) {
-      return [repackRuntimePlugin, ...runtimePlugins];
-    }
-
-    return runtimePlugins;
-  }
-
-  private getModuleFederationPlugin(compiler: Compiler) {
-    if (isRspackCompiler(compiler)) {
-      return require('@module-federation/enhanced/rspack')
-        .ModuleFederationPlugin;
-    }
-    return require('@module-federation/enhanced/webpack')
-      .ModuleFederationPlugin;
-  }
-
-  private getDefaultSharedDependencies() {
-    return {
-      react: { singleton: true, eager: true },
-      'react-native': { singleton: true, eager: true },
-    };
-  }
-
-  /**
-   * As including 'react-native' as a shared dependency is not enough to support
-   * deep imports from 'react-native' (e.g. 'react-native/Libraries/Utilities/PixelRatio'),
-   * we need to add deep imports using an undocumented feature of ModuleFederationPlugin.
-   *
-   * When a dependency has a trailing slash, deep imports of that dependency will be correctly
-   * resolved by reaching out to the shared scope. This also ensures single instances of things
-   * like 'assetsRegistry'. Additionally, we mark every package from '@react-native' group as shared
-   * as well, as these are used by React Native too.
-   *
-   * Reference: https://stackoverflow.com/questions/65636979/wp5-module-federation-sharing-deep-imports
-   * Reference: https://github.com/webpack/webpack/blob/main/lib/sharing/resolveMatchedConfigs.js#L77-L79
-   *
-   * @param shared shared dependencies configuration from ModuleFederationPlugin
-   * @returns adjusted shared dependencies configuration
-   *
-   * @internal
-   */
-  private adaptSharedDependencies(shared: MF.Shared): MF.Shared {
-    const sharedDependencyConfig = (eager?: boolean) => ({
-      singleton: true,
-      eager: eager ?? true,
-      requiredVersion: '*',
-    });
-
-    const findSharedDependency = (
-      name: string,
-      dependencies: MF.Shared
-    ): MF.SharedConfig | string | undefined => {
-      if (Array.isArray(dependencies)) {
-        return dependencies.find((item) =>
-          typeof item === 'string' ? item === name : Boolean(item[name])
-        );
-      }
-      return dependencies[name];
-    };
-
-    const sharedReactNative = findSharedDependency('react-native', shared);
-    const reactNativeEager =
-      typeof sharedReactNative === 'object'
-        ? sharedReactNative.eager
-        : undefined;
-
-    if (!this.deepImports || !sharedReactNative) {
-      return shared;
-    }
-
-    if (Array.isArray(shared)) {
-      const adjustedSharedDependencies = [...shared];
-      if (!findSharedDependency('react-native/', shared)) {
-        adjustedSharedDependencies.push({
-          'react-native/': sharedDependencyConfig(reactNativeEager),
-        });
-      }
-      if (!findSharedDependency('@react-native/', shared)) {
-        adjustedSharedDependencies.push({
-          '@react-native/': sharedDependencyConfig(reactNativeEager),
-        });
-      }
-      return adjustedSharedDependencies;
-    }
-    const adjustedSharedDependencies = { ...shared };
-    if (!findSharedDependency('react-native/', shared)) {
-      Object.assign(adjustedSharedDependencies, {
-        'react-native/': sharedDependencyConfig(reactNativeEager),
-      });
-    }
-    if (!findSharedDependency('@react-native/', shared)) {
-      Object.assign(adjustedSharedDependencies, {
-        '@react-native/': sharedDependencyConfig(reactNativeEager),
-      });
-    }
-    return adjustedSharedDependencies;
+  constructor(config: ModuleFederationPluginV1Config) {
+    this.plugin = new ModuleFederationPluginV1(config);
   }
 
   apply(compiler: Compiler) {
-    this.ensureModuleFederationPackageInstalled(compiler.context);
+    const logger = compiler.getInfrastructureLogger('ModuleFederationPlugin');
 
-    // MF2 produces warning about not supporting async await
-    // we can silence this warning since it works just fine
-    compiler.options.ignoreWarnings = compiler.options.ignoreWarnings ?? [];
-    compiler.options.ignoreWarnings.push(
-      (warning) => warning.name === 'EnvironmentNotSupportAsyncWarning'
+    logger.warn(
+      'Notice: ModuleFederationPlugin currently points to ModuleFederationPluginV1. ' +
+        'Re.Pack 5 introduced ModuleFederationPluginV2, which addresses many previous limitations. ' +
+        'In the next major version of Re.Pack, ModuleFederationPlugin will point to ModuleFederationPluginV2. ' +
+        'We recommend switching to the new ModuleFederationPluginV2 by importing it directly. ' +
+        'If you want to keep using ModuleFederationPluginV1, which is no longer being iterated on, ' +
+        'you can import ModuleFederationPluginV1 directly to prevent this warning from being shown every time.'
     );
 
-    const ModuleFederationPlugin = this.getModuleFederationPlugin(compiler);
-
-    const libraryConfig = this.config.exposes
-      ? {
-          name: this.config.name,
-          type: 'self',
-          ...this.config.library,
-        }
-      : undefined;
-
-    const sharedConfig = this.adaptSharedDependencies(
-      this.config.shared ?? this.getDefaultSharedDependencies()
-    );
-
-    const shareStrategyConfig = this.config.shareStrategy ?? 'loaded-first';
-
-    const runtimePluginsConfig = this.adaptRuntimePlugins(
-      compiler.context,
-      this.config.runtimePlugins
-    );
-
-    const config: MF.ModuleFederationPluginOptions = {
-      ...this.config,
-      library: libraryConfig,
-      shared: sharedConfig,
-      shareStrategy: shareStrategyConfig,
-      runtimePlugins: runtimePluginsConfig,
-    };
-
-    new ModuleFederationPlugin(config).apply(compiler);
+    this.plugin.apply(compiler);
   }
 }
