+++ packages/repack/src/plugins/index.ts
@@ -5,5 +5,6 @@ export * from './OutputPlugin';
 export * from './RepackTargetPlugin';
 export * from './ModuleFederationPlugin';
 export * from './ModuleFederationPluginV1';
+export * from './ModuleFederationPluginV2';
 export * from './CodeSigningPlugin';
 export * from './ChunksToHermesBytecodePlugin';
